Anatomical measurements

This example dataset includes ROXAS output files (cf. von Arx & Carrer, 2014) with lumen area and cell wall thickness measurements from; i) Norway Spruce (Picea abies) from the Loetschental, Switzerland (2007-2013), ii) Siberian larch (Larix siberica) from Siberia, Russia (2007-2010), iii) Stone pine (Pinus cembra) from the Dolomite mountains, Italy (2007-2010), and iv) Scots pine (Pinus sylvestris) from the eastern lowlands, Germany (2007-2010).

anatomy.data

anatomy.data

Format

Provides a data.frame with 18838 rows and 7 columns containing data on:

[ ,1] ID = Site and species id <factor>
[ ,2] CID = Unique cell id <integer>
[ ,3] YEAR = Year of the ring <integer>
[ ,4] CA = Lumen size (micron) <numeric>
[ ,5] XCAL = X-value of cell center in calibrated coordinate system (micron) <numeric>
[ ,6] YCAL = Y-value of cell center in calibrated coordinate system (micron) <numeric>
[ ,7] CWTALL = Mean thickness of all cell walls (micron) <numeric>

References

von Arx, G., & Carrer, M. (2014) ROXAS - A new tool to build centuries-long tracheid-lumen chronologies in conifers. Dendrochronologia 32, 290-293.

Examples

#viewing anatomy data
head(anatomy.data)

#>          ID CID YEAR   CA  XCAL   YCAL CWTALL
#> 1 LOT_PICEA  31 2007 1252 -1481 108980   3.01
#> 2 LOT_PICEA  32 2007 1136 -1440 108979   2.76
#> 3 LOT_PICEA  34 2007 1139 -1402 108979   2.04
#> 4 LOT_PICEA  40 2007 1320 -1365 109000   2.69
#> 5 LOT_PICEA  45 2007 1688 -1317 109006   2.36
#> 6 LOT_PICEA  50 2007  340 -1288 109005   2.21